refactor add docs and change modules name
